[PLAT-726 C-2246] Add new notification flow
