+++ packages/web/src/common/store/notifications/constants.ts
@@ -1,0 +1,1 @@
+export const NOTIFICATION_LIMIT_DEFAULT = 20
